addons-react: extract FluxibleProvider from FluxibleContext module
