+++ packages/fluxible-addons-react/src/index.js
@@ -3,6 +3,7 @@
  * Copyrights licensed under the New BSD License. See the accompanying LICENSE file for terms.
  */
 export { default as FluxibleComponent } from './FluxibleComponent';
+export { default as FluxibleProvider } from './FluxibleProvider';
 export { default as batchedUpdatePlugin } from './batchedUpdatePlugin';
 export { default as connectToStores } from './connectToStores';
 export { default as createElementWithContext } from './createElementWithContext';
